move dir package
